requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backoff==2.2.1
binaryornot==0.4.4
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cookiecutter==2.6.0
debugpy==1.8.12
decorator==5.1.1
docker==7.1.0
earthengine-api==1.5.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
filelock==3.17.0
fire==0.7.0
frozenlist==1.5.0
fsspec==2025.2.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
lightning==2.5.0.post0
lightning-utilities==0.12.0
lightning_sdk==0.1.55
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyreadstat==1.2.8
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.5.2
simple-term-menu==1.6.6
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
text-unidecode==1.3
timm==1.0.14
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchmetrics==1.6.1
torchvision==0.21.0+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
yarl==1.18.3
